Move `validateAndInitializeFirebaseWithConfig` function to a separate file
